Refactorización de la vista pokemon
