+++ lib/widgets/SecondaryCard/secondary_card.dart
@@ -1,9 +1,10 @@
 
 import 'package:flutter/material.dart';
+import 'package:mobile_pokedex/widgets/SecondaryCard/BodyMeasurement%20Value/body%20_measurement%20_value.dart';
 import 'package:string_capitalize/string_capitalize.dart';
 
 import '../StatInformation/stat_information.dart';
-import '../TypeCard/type_card.dart';
+import 'TypeCard/type_card.dart';
 
 class SecondaryCard extends StatelessWidget {
   const SecondaryCard({
@@ -53,53 +54,18 @@ class SecondaryCard extends StatelessWidget {
             Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                Column(
-                  children: [
-                    Text("$weight Kg", 
-                      style: TextStyle(
-                        fontSize: 18
-                      ),
-                    ),
-                    Text("Weigth", 
-                      style: TextStyle(
-                        fontWeight: FontWeight.bold,
-                        fontSize: 22
-                      ),
-                    )
-                  ],
-                ),
+                BodyMeasurementValue(bodyMeasurement: "Weight", value: "$weight Kg"),
                 SizedBox(width: 20,),
-                Column(
-                  children: [
-                    Text("$height m",
-                      style: TextStyle(
-                        fontSize: 18
-                      ),
-                    ),
-                    Text("Heigth",
-                      style: TextStyle(
-                        fontWeight: FontWeight.bold,
-                        fontSize: 22
-                      ),
-                    )
-                  ],
-                )
+                BodyMeasurementValue(bodyMeasurement: "Height", value: "$height m",)
               ],
             ),
-            Container(
-              decoration: BoxDecoration(
-                color: Colors.black12,
-                borderRadius: BorderRadius.all(Radius.circular(12))
-              ),
-              margin: EdgeInsets.fromLTRB(40,10,40,0),
-              padding: EdgeInsets.only(bottom: 10),
-              child:Column(
-                children: stats.map((stat) => StatInformation(stat: stat)).toList()
-              ), 
-            )
+            StatInformation(stats: stats)
           ],
         ),
       ),
     );
   }
 }
+
+
+
