+++ lib/widgets/StatInformation/stat_information.dart
@@ -1,69 +1,27 @@
 import 'package:flutter/material.dart';
-import 'package:mobile_pokedex/models/pokemon_model.dart';
-import 'package:string_capitalize/string_capitalize.dart';
 
-class StatInformation extends StatelessWidget{
+import 'StatValue/stat_value.dart';
 
-  StatInformation({
+class StatInformation extends StatelessWidget {
+  const StatInformation({
     super.key,
-    required this.stat
+    required this.stats,
   });
 
-  final Stat stat;
-  final Map<String, int> color ={
-    'hp': 0xFFB6DB1D,
-    'attack': 0xFFDB4E1D,
-    'defense': 0xFF2A6BDB,
-    'special Attack': 0xFFA656DC,
-    'special Defense': 0xFFA656DC,
-    'speed': 0xFFFFC107,
-  };
-
-  Color getColorByStat( String type){
-    return Color(color[type] ?? 0xFFFFC107);
-  }
-
-
-  double statPercentageCalculator(int statValue){
-    double unroundedResult = statValue/300;
-    return double.parse(unroundedResult.toStringAsFixed(2));
-  }
+  final List<dynamic> stats;
 
   @override
-  Widget build(BuildContext context){
-    return Column(
-      children: [
-        Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Text(stat.name.capitalize(), 
-              style: TextStyle(
-                fontWeight: FontWeight.bold,
-                fontSize: 22,
-              ) ,
-            ),
-            Text(": ${stat.baseValue}",
-              style: TextStyle(
-                fontSize: 20
-              ),
-            )
-          ],
-        ),
-        Container(
-          width: 250,
-          height: 12,
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.all(Radius.circular(10)),
-            border: Border.all(width: 2)
-          ),
-          child: LinearProgressIndicator(
-            value: statPercentageCalculator(stat.baseValue),
-            backgroundColor: Colors.grey,
-            valueColor: AlwaysStoppedAnimation(getColorByStat(stat.name)),
-            borderRadius: BorderRadius.all(Radius.circular(10)),
-          ),
-        )
-      ],
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+        color: Colors.black12,
+        borderRadius: BorderRadius.all(Radius.circular(12))
+      ),
+      margin: EdgeInsets.fromLTRB(40,10,40,0),
+      padding: EdgeInsets.only(bottom: 10),
+      child:Column(
+        children: stats.map((stat) => StatValue(stat: stat)).toList()
+      ), 
     );
   }
 }
